+++ utils/compute-cmvn-stats.py
@@ -8,7 +8,7 @@ import numpy as np
 from espnet.utils.cli_utils import FileReaderWrapper
 from espnet.utils.cli_utils import FileWriterWrapper
 from espnet.utils.cli_utils import get_commandline_args
-from espnet.utils.io_utils import Preprocessing
+from espnet.transform.transformation import Transformation
 
 
 def main():
@@ -87,7 +87,7 @@ def main():
             args.out_filetype = 'npy'
 
     if args.preprocess_conf is not None:
-        preprocessing = Preprocessing(args.preprocess_conf)
+        preprocessing = Transformation(args.preprocess_conf)
         logging.info('Apply preprocessing: {}'.format(preprocessing))
     else:
         preprocessing = None
